REFACTOR
NEW/public/shared/data/config.json:
  * Better separated data.

NEW/public/shared/data/regenerate.js:
  * Generates tileIdsByTilename.json
  * Generates tilenamesByIndex.json
  * Generates coordsByIndex.json
  * Generates indexByCoords.json
  * Based on data from: config.json, tileCoords.json

NEW/backend/python/:
  * server.py
    * Main program.
  * class_battery.py
    * Handles the INA219 battery module.
  * class_webServer.py
    * HTTP web server for testing purposes.
  * class_websocketServer.py
    * WebSocket server to handle requests from Node.
  * class_graphics.py
    * TODO: Handles drawing.

NEW/backend/node/:
  * Server can start.
  * All config files are loaded.
  * Old processes can be removed.
  * Draw functions are ready to test. TODO: test.
  * GPIO functions are ready to test. TODO: test.
