--- backend/config.json
@@ -1,56 +1,0 @@
-{
-	"app": {
-		"hide_APP_consolelog": false
-	},
-	"server":{
-		"host": "0.0.0.0",
-		"port": 7777
-	},
-	"gpio":{
-		"active": true
-	},
-	"ws":{
-		"active": true
-	},
-	"lcd":{
-		"active": true,
-		"fps":5,
-		"width"       : 240,
-		"height"      : 240,
-		"activeTileset": "tileset_8x8",
-		"tilesets":{
-			"tileset_8x8": {
-				"file": "tileset_8x8.png",
-				"n":{
-					"tileWidth"   : 8, 
-					"tileHeight"  : 8 
-				},
-				"s":{
-					"tileWidth"   : 8, 
-					"tileHeight"  : 8 
-				}
-			},
-			"tileset_12x12": {
-				"file": "tileset_12x12.png",
-				"n":{
-					"tileWidth"   : 12, 
-					"tileHeight"  : 12 
-				},
-				"s":{
-					"tileWidth"   : 12, 
-					"tileHeight"  : 12 
-				}
-			}
-		}
-	},
-	"python":{
-		"active": true,
-		"serverType": "ws",
-		"ws":{
-			"port": 7778
-		},
-		"http":{
-			"port": 7778
-		}
-	}
-}
