+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test3",
+  "name": "commander_mini",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
