--- public/index.css
@@ -1,164 +1,0 @@
-* {
-	box-sizing: border-box;
-}
-
-/* BODY */
-body {
-	background-color: darkblue;
-}
-
-/* SHARED */
-.inline-block {
-	display: inline-block;
-	vertical-align: top;
-}
-.bg1{
-	background-color:gray;
-}
-.center{
-	text-align:center;
-}
-
-/* BUTTONS */
-.button1 {
-	min-width: 32px;
-}
-.button2 {
-	width: 100%;
-}
-.button3 {
-	width: 300px;
-	text-align: left;
-	font-family: monospace;
-}
-.button1:hover,
-.button2:hover,
-.button3:hover {
-	background-color:goldenrod;
-}
-.button1:active,
-.button2:active,
-.button3:active {
-	background-color:yellow;
-}
-
-/* DIV CONTAINERS */
-#container1,
-#liveControls,
-#canvasDisplay,
-#debugOutput
-#debugControls{
-	display: inline-block;
-	vertical-align: top;
-	padding: 2px;
-	background-color: black;
-	border-radius: 5px;
-}
-#container1{
-	border: 5px solid navajowhite;
-}
-#liveControls{
-}
-#debugControls {
-	border: 5px solid navajowhite;
-}
-#debugOutput {
-	border: 5px solid navajowhite;
-	width: 300px;
-	min-height: 275px;
-}
-#canvasDisplay {
-	width: 360px;
-}
-#debugOutput_output{
-	width : 100%;
-	height: 220px;
-	font-family:monospace;
-	white-space:pre;
-}
-.disconnected{
-	border-color:red !important;
-}
-
-/* CANVAS */
-#CANVAS1 {
-	background-color: black;
-	width: 100%;
-	height: 100%;
-}
-
-/* TABLES */
-.tableType1 {}
-.tableType1 {
-	margin-bottom: 5px;
-	border: 2px solid black;
-	border-top: 0px;
-	border-radius: 0px 0px 5px 5px;
-	border-collapse: collapse;
-	background-color: darkkhaki;
-	width:100%;
-}
-.tableType1 caption {
-	background-color: #c1c1c1;
-	font-weight: bold;
-	text-decoration: underline;
-	border: 2px solid black;
-	border-bottom: 1px solid black;
-	border-radius: 5px 5px 0px 0px;
-	padding: 4px
-}
-.tableType1 th,
-.tableType1 td {
-	padding: 4px;
-}
-
-.tableType1 th {}
-
-#vramtable{
-    border: 3px solid dodgerblue;
-    /* width: 250px; */
-    background-color: black;
-    border-collapse: collapse;
-}
-#vramtable td{
-    border: 1px solid;
-    border-color: blue;
-    padding: 0px;
-    margin: 0px;
-    /* white-space: pre-wrap; */
-    background-color: #1c1a1a;
-    width: 0px;
-    height: 0px;
-    border-spacing: 0px;
-    text-indent: 0px;
-    white-space: pre;
-}
-#vramtable th{}
-#vramtable caption {
-	background-color: #c1c1c1;
-	font-weight: bold;
-	text-decoration: underline;
-	border: 2px solid black;
-	border-bottom: 1px solid black;
-	border-radius: 5px 5px 0px 0px;
-	padding: 4px
-}
-.vram_smallfont{
-	/* font-family:
-	monospace; */
-	/* font-size: 5px; */
-}
-.tileCanvas{
-	background-color:
-	black;
-	/* border:1px solid white; */
-	/* display: inline-block; */
-	margin: 1px;
-	width: 8px;
-	height: 8px;
-}
-#fps2_curFrame{
-	color: darkblue;
-	font-size: 10px;
-	background-color: darkgoldenrod;
-}
