+++ public/shared/README.md
@@ -1,0 +1,47 @@
+# /SHARED
+- These files govern the tileset image, tiles, tile dimensions, lcd dimensions, and display framerate.
+
+# DEFINITIONS:
+- VRAM tile index   
+  - The tile on the screen indicated by x,y.
+- VRAM 1D tile index
+  - VRAM start index for x,y. Has 3 values for each x.
+- Tile coords       
+  - x and y coords of a tile relative to it's tileset image and the tile dimensions. 
+  - x and y need to be multiplied by tileWidth and tileHeight respectively.  
+
+# FILES (Required by this script):
+tileCoords.json
+- Used to determine the region of the tileset image is used by the named tile.
+  - Object of key/value pairs. 
+  - Key is the tile name (as string).
+  - NOTE: Numbers and space are prefixed with "n".
+  - Value is object having keys "L" and "T".
+    - "L" and "T" indicate how many tiles from the left or the top of the image to get the starting point of the tile.
+
+config.json
+  - Keeps settings for the app.
+  - Key "lcd" contains the data for the tiles, tileset image, and dimensions.
+
+# FILES (generated by this script):
+coordsByIndex.json
+- Provide the VRAM tile index to get the x, y coordinates.
+  - 2D array. Each y row consists of x arrays each having:
+    - index 0: y tile coord.
+	- index 1: x tile coord.
+	- index 2: VRAM 1D array start index index. 
+	- index 3: VRAM tile index based on y and x. 
+      - Example: if there are 30 rows and 30 columns then there are 900 VRAM tile index.
+      - Example: if there are 30 rows and 30 columns then there are 2700 VRAM 1D indexes.
+
+indexByCoords.json
+- Used to determine the VRAM 1D array index by providing the y and x coordinate.
+  - Object of key/value pairs. 
+  - Key is the index (as string).
+  - Value is the x,y coordinate in the form of [x, y].
+
+tileIdsByTilename.json
+- tileIdsByTilename
+
+tilenamesByIndex.json
+- tilenamesByIndex
