+++ public/shared/tileCoords.json
