+++ public/shared/tileIdsByTilename.json
@@ -1,0 +1,85 @@
+{
+ "n ": 0,
+ "!": 1,
+ "\"": 2,
+ "#": 3,
+ "$": 4,
+ "%": 5,
+ "&": 6,
+ "'": 7,
+ "(": 8,
+ ")": 9,
+ "*": 10,
+ "+": 11,
+ ",": 12,
+ "-": 13,
+ ".": 14,
+ "/": 15,
+ "n0": 16,
+ "n1": 17,
+ "n2": 18,
+ "n3": 19,
+ "n4": 20,
+ "n5": 21,
+ "n6": 22,
+ "n7": 23,
+ "n8": 24,
+ "n9": 25,
+ ":": 26,
+ ";": 27,
+ "<": 28,
+ "=": 29,
+ ">": 30,
+ "?": 31,
+ "@": 32,
+ "A": 33,
+ "B": 34,
+ "C": 35,
+ "D": 36,
+ "E": 37,
+ "F": 38,
+ "G": 39,
+ "H": 40,
+ "I": 41,
+ "J": 42,
+ "K": 43,
+ "L": 44,
+ "M": 45,
+ "N": 46,
+ "O": 47,
+ "P": 48,
+ "Q": 49,
+ "R": 50,
+ "S": 51,
+ "T": 52,
+ "U": 53,
+ "V": 54,
+ "W": 55,
+ "X": 56,
+ "Y": 57,
+ "Z": 58,
+ "[": 59,
+ "\\": 60,
+ "]": 61,
+ "^": 62,
+ "_": 63,
+ "tile1": 64,
+ "tile2": 65,
+ "tile3": 66,
+ "tile4": 67,
+ "tile_red": 68,
+ "tile_green": 69,
+ "tile_blue": 70,
+ "cursor1": 71,
+ "cursor2": 72,
+ "cursor3": 73,
+ "cursor4": 74,
+ "nochar": 75,
+ "clock1": 76,
+ "battcharge1": 77,
+ "battcharge2": 78,
+ "batt1": 79,
+ "batt2": 80,
+ "batt3": 81,
+ "batt4": 82
+}
