+++ public/shared/tilenamesByIndex.json
@@ -1,0 +1,85 @@
+[
+ "n ",
+ "!",
+ "\"",
+ "#",
+ "$",
+ "%",
+ "&",
+ "'",
+ "(",
+ ")",
+ "*",
+ "+",
+ ",",
+ "-",
+ ".",
+ "/",
+ "n0",
+ "n1",
+ "n2",
+ "n3",
+ "n4",
+ "n5",
+ "n6",
+ "n7",
+ "n8",
+ "n9",
+ ":",
+ ";",
+ "<",
+ "=",
+ ">",
+ "?",
+ "@",
+ "A",
+ "B",
+ "C",
+ "D",
+ "E",
+ "F",
+ "G",
+ "H",
+ "I",
+ "J",
+ "K",
+ "L",
+ "M",
+ "N",
+ "O",
+ "P",
+ "Q",
+ "R",
+ "S",
+ "T",
+ "U",
+ "V",
+ "W",
+ "X",
+ "Y",
+ "Z",
+ "[",
+ "\\",
+ "]",
+ "^",
+ "_",
+ "tile1",
+ "tile2",
+ "tile3",
+ "tile4",
+ "tile_red",
+ "tile_green",
+ "tile_blue",
+ "cursor1",
+ "cursor2",
+ "cursor3",
+ "cursor4",
+ "nochar",
+ "clock1",
+ "battcharge1",
+ "battcharge2",
+ "batt1",
+ "batt2",
+ "batt3",
+ "batt4"
+]
