+++ public/shared/tileset_8x8.png
